requirements.txt
# HTTP & Crawling
httpx==0.27.0
feedparser==6.0.11
beautifulsoup4==4.12.3

# Graph DB
neo4j==5.20.0

# NLP — SPECTER2 paper-level embeddings (Phase 2)
transformers==4.41.0
adapters==0.2.1
torch==2.3.0
numpy==1.26.4

# Topic clustering (Phase 3)
umap-learn==0.5.6
hdbscan==0.8.38
scikit-learn==1.5.0

# LLM (fallback judge + on-demand summaries)
together==1.2.1

# Anthropic (optional, for summaries)
anthropic==0.28.0

# CLI
typer==0.12.3
rich==13.7.1

# Config
python-dotenv==1.0.1
pydantic==2.7.3
pydantic-settings==2.3.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
